Add CourseDetails folder
