+++ client/src/components/CourseCard.tsx
@@ -1,29 +1,28 @@
 import React from 'react';
 
-const CourseCard = ({ image, title, text }) => {
+const CourseCard = ({ id, image, title, text }) => {
   return (
-    <div>
-      <div className="max-w-sm rounded overflow-hidden shadow-lg">
-        <img className="w-full" src={image} alt="image" />
+    <div className="flex flex-wrap bg-gray-200 shadow-lg">
+      <div className="w-60">
+        <div className="">
+          <img
+            className="w-full object-cover h-24 aspect-w-1 aspect-h-1 p-2"
+            src={image}
+            alt="image"
+          />
+        </div>
         <div className="px-6 py-4">
           <div className="font-bold text-xl mb-2">{title}</div>
-          <p className="text-gray-700 text-base">
-          {text}
+          <p className="min-h-40 overflow-hidden text-gray-700 text-base">
+            {text}
           </p>
         </div>
-        <div className="px-6 pt-4 pb-2">
-          <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">
-            #photography
-          </span>
-          <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">
-            #travel
-          </span>
-          <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">
-            #winter
-          </span>
+        <div className="flex justify-end text-blue-500 px-6 pt-4 pb-2">
+          <button>
+            <a href={`/details/${id}`}>details</a>
+          </button>
         </div>
       </div>
-
     </div>
   );
 };
