+++ client/src/components/courseData.js
@@ -1,0 +1,32 @@
+const data = [
+  {
+    id: '1',
+    title: 'HTML basics',
+    image:
+      'https://www.udacity.com/blog/wp-content/uploads/2020/06/HTML_Blog-scaled.jpeg',
+    text: 'Learn the basics of HTML.',
+  },
+  {
+    id: '2',
+    title: 'CSS basics',
+    image:
+      'https://cdn-media-2.freecodecamp.org/w1280/5f9c9d7b740569d1a4ca3800.jpg',
+    text: 'Fundamental of UX for Application design The automated process all your website tasks. Discover tools and techniques to engage effectively with vulnerable children and young people.',
+  },
+  {
+    id: '3',
+    title: 'Javascript basics',
+    image:
+      'https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fthumbs.dreamstime.com%2Fb%2Fcoding-concept-html-code-coding-programming-concept-html-source-code-tilted-vector-illustration-155503904.jpg&f=1&nofb=1&ipt=3ac8ffd453ffda7c0ddc7f218474984acc961fa83fd356d3a181fd50c58c60db&ipo=images',
+    text: 'Fundamental of UX for Application design The automated process all your website tasks. Discover tools and techniques to engage effectively with vulnerable children and young people.',
+  },
+  {
+    id: '4',
+    title: 'React basics',
+    image:
+      'https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fcdn.filestackcontent.com%2FG33P8iuoRuinYep5O6dt&f=1&nofb=1&ipt=b8491976fea8b591ced9d25cb2b4cd003adeecda07631028e3013f41276f28a9&ipo=images',
+    text: 'Fundamental of UX for Application design The automated process all your website tasks. Discover tools and techniques to engage effectively with vulnerable children and young people.',
+  },
+];
+
+export default data;
