--- client/src/lib/utils.ts
@@ -1,7 +1,0 @@
-// component for chadcn ui library - can be deleted
-import { type ClassValue, clsx } from "clsx"
-import { twMerge } from "tailwind-merge"
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
-}
